requirements.txt
aioredis==1.3.1
amqp==5.0.9
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.2.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.10
django-cors-headers==3.10.1
django-countries==7.2.1
django-environ==0.8.1
django-phonenumber-field==6.0.0
django-rest-swagger==2.2.0
djangorestframework==3.13.0
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
msgpack==1.0.3
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.12.39
pika==1.2.0
pika-stubs==0.1.3
Pillow==8.4.0
priority==1.3.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-decouple==3.5
pytz==2021.3
redis==4.1.0
requests==2.26.0
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zope.interface==5.4.0
